Updating series message processing for flaired posts
This change makes it so that when posting the RemindMeBot comment on posts,
previous posts are reprocessed just in case they were flaired as series and
the bot didn't detect the post was a series when it originally processed it.

The bot will not determine whether or not previously seen posts have been
flaired after the fact, and will post the series message if that's the case.