Switch from RemindMeBot to UpdateMeBot
I run the UpdateMeBot that allows people to subscribe to an author in a subreddit and get reminders each time they post. Users can subscribe via PM, but they can also subscribe by posting a public comment with the phrase "SubscribeMe" or "UpdateMe". The method I use to find comments is occasionally a bit flaky, so I try to encourage users to use a PM instead. I've noticed a relatively large number of public comments in the /r/nosleep subreddit, often in reply to this bots stickied comment. So I figured I would make this pull request to change the link over from the RemindMeBot to my bot.