Manually encode url
I've been getting some messages from users that the spaces in the message link are showing up as "+"s. Normally it's correct behavior to encode url parameters like this, but it looks like it's breaking on the android mobile site and not unencoding them properly. Python 3 has a way to better control the functionality of urlencode, but since this is python 2, the simplest way is to do it manually. Valid usernames can't contain any characters that need encoding, so this should be safe.