Migrate to pydantic for settings
The old configuration system had some features that have never been used
for actually running the bot in production, like running from an .ini
config file or using a cascade of overrides.

This change gets rid of the configuration handling that was built in to
the code for the past couple years, so it scraps the .ini file support,
manually reading from the environment, and all the overriding support.

Instead, pydantic has been introduced, and the configuration was broken
out into a settings type rather than a bare dict. This also adds support
for .env and more flexibility for handling stuff like redis URLs.