Fix tests, create PostAnalyzer
Combined all the post analysis stuff for things like tags and formatting
into a single class, PostAnalyzer, that also produces a single dataclass
which stores all the relevant information. This allowed a big simplification
of the various checks the bot did to determine if a post was good or not.