Account for listings where user deleted own post
By using 'new()' listings, the bot uses the 'before' parameter
on the listing to get posts newer than the last one it's seen. It
already had code handling the situation where it was 'removed', but
that only applies to moderator/bot deletions, not deletions by the
user. This adds checks to see if the latest post has a null author
or isn't robot indexable, which means it's been removed.