Skip time enforcement if user removed own post
When users delete their own post, they shouldn't have a
last activity check for time limit enforcement. In the
previous versions where it polled using /search, this was
never a problem because it was freshly looking up posts
every time it processed posts. Here, though, we need to
check if the last activity by a user is a post that is still
active/undeleted.