Cache post ID and not post name for activity
This is to fix issue #123, where time limit enforcement was broken because
we're storing the fully qualified post ID, and not just the base36 ID of
posts. This breaks submission deletion check when the bot is trying to
determine if a post has been deleted or not, and it would return 'true' in
the cases where an author was trying to post within the timelimit and they
had cached activity.